Moved around on some startupservices, and created a host.test project to have hosting tests there
