+++ MigrationTools.sln
@@ -107,6 +107,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MigrationTools.Clients.InMe
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MigrationTools.Clients.FileSystem.Tests", "src\MigrationTools.Clients.FileSystem.Tests\MigrationTools.Clients.FileSystem.Tests.csproj", "{3191B594-1C3A-4245-8AFE-076EA142B778}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MigrationTools.Host.Tests", "src\MigrationTools.Host.Tests\MigrationTools.Host.Tests.csproj", "{6F51CE1D-FEBD-4F7F-A462-27A4745924F4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -165,6 +167,10 @@ Global
 		{3191B594-1C3A-4245-8AFE-076EA142B778}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3191B594-1C3A-4245-8AFE-076EA142B778}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3191B594-1C3A-4245-8AFE-076EA142B778}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6F51CE1D-FEBD-4F7F-A462-27A4745924F4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6F51CE1D-FEBD-4F7F-A462-27A4745924F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6F51CE1D-FEBD-4F7F-A462-27A4745924F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6F51CE1D-FEBD-4F7F-A462-27A4745924F4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -190,6 +196,7 @@ Global
 		{54DF41E6-9F94-42DD-877E-437DC6E9F3E2} = {13FE9FE2-9C6E-4866-BBAD-106F5BF765F1}
 		{504D6953-E83D-498D-A1C7-E511F14B9563} = {13FE9FE2-9C6E-4866-BBAD-106F5BF765F1}
 		{3191B594-1C3A-4245-8AFE-076EA142B778} = {13FE9FE2-9C6E-4866-BBAD-106F5BF765F1}
+		{6F51CE1D-FEBD-4F7F-A462-27A4745924F4} = {5C05DA7B-B70F-4A87-920B-DCD14EA7670D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {62EE0B27-C55A-46EE-8D17-1691DE9BBD50}
