+++ src/MigrationTools.Host.Tests/FakeMigrationClientConfig.cs
@@ -1,0 +1,17 @@
+﻿using MigrationTools.Configuration;
+
+namespace MigrationTools.Tests
+{
+    public class FakeMigrationClientConfig : IMigrationClientConfig
+    {
+        public IMigrationClientConfig PopulateWithDefault()
+        {
+            return this;
+        }
+
+        public override string ToString()
+        {
+            return "FakeMigration";
+        }
+    }
+}
