+++ src/MigrationTools.Host/MigrationToolHost.cs
@@ -11,7 +11,7 @@ using MigrationTools.CommandLine;
 using MigrationTools.Configuration;
 using MigrationTools.Engine.Containers;
 using MigrationTools.Host.CustomDiagnostics;
-using MigrationTools.Services;
+using MigrationTools.Host.Services;
 using Serilog;
 using Serilog.Core;
 using Serilog.Events;
