--- src/MigrationTools.Tests/Core/Configuration/EngineConfigurationBuilderStub.cs
@@ -1,34 +1,0 @@
-﻿using Microsoft.Extensions.Logging;
-
-namespace MigrationTools.Configuration.Tests
-{
-    public class EngineConfigurationBuilderStub : IEngineConfigurationBuilder
-    {
-        private IEngineConfigurationBuilder ecb;
-
-        public EngineConfigurationBuilderStub(ILogger<EngineConfigurationBuilder> logger)
-        {
-            ecb = new EngineConfigurationBuilder(logger);
-        }
-
-        public EngineConfiguration BuildDefault()
-        {
-            return ecb.BuildDefault();
-        }
-
-        public EngineConfiguration BuildFromFile(string configFile = "configuration.json")
-        {
-            return ecb.BuildDefault();
-        }
-
-        public EngineConfiguration BuildWorkItemMigration()
-        {
-            return ecb.BuildWorkItemMigration();
-        }
-
-        public EngineConfiguration CreateEmptyConfig()
-        {
-            return ecb.CreateEmptyConfig();
-        }
-    }
-}
