--- src/MigrationTools/CommandLine/Execute/ExecuteCommand.cs
@@ -1,6 +1,0 @@
-﻿namespace MigrationTools.CommandLine.Execute
-{
-    internal class ExecuteCommand
-    {
-    }
-}
